Fix-95 : Provisioner approach to restart docker
* Fix-95: TLS certificates are not valid for private network IP
* Reverted to provisioner approach due to adb-atomic-developer-bundle/issues/283
