+++ Vagrantfile
@@ -13,9 +13,6 @@ Vagrant.configure(2) do |config|
 
   config.vm.box = "projectatomic/adb"
 
-  # This is the default setup
-  # config.servicemanager.services = 'docker'
-
-  # Enable multiple services as comma separated list.
-  # config.servicemanager.services = 'docker, openshift'
+  # enable service-manager plugin
+  config.vm.provision :servicemanager
 end
